IJT-2393
If OIDC token expires (1 hr), it turns nil.  The `mgr.validateToken()` throws a `"Could not parse the given JWT string payload."` error when it tries to validate the nil token.  Rather than check it, automatically return nil if the token is nil so it will get regenerated from refresh token in `getUser()`